ast: Collapse `AssignExpr` into more generic `ValueExpr`
The primary goal of this change is to reduce the amount of indirection
and node unwrapping needed to retrieve the final node data.

An `AssignExpr`  is unnecessary as there is not really any extra context
needed in the context of assignment. The only semantic that matters is
that it should yield a value (counting things like references too,
which is valid not just in assignments).