ast: Switch to arena allocation for `NodeData`
Having the inconsistency of using `SmallVec` and `Vec` in some places
bothered me when I was using an arena allocator for everything else.

The reason I had originally choosen to do it that way was due to possible
lifetime pollution for small arena collections in a node data
(e.g., the parameters for a `FnCallExpr`); however, I realized I can
simply maintain a global list of these indices and store a span in the node.

* Removed dependency on `smallvec`, for now
* Cargo fmt